+++ eden/week6/src/modules/missions/missions.routes.ts
@@ -1,9 +1,16 @@
 import { Router } from "express";
-import { handleAddMission } from "./controllers/mission.controller.js";
+import { handleAddMission, handleChallengeMission, handleListStoreMissions, handleListUserMissions } from "./controllers/mission.controller.js";
 
 const missionsRouter = Router();
 
-// /api/v1/stores/:storeId/missions
-missionsRouter.post("/stores/:storeId/missions", handleAddMission);
+
+//미션 추가
+missionsRouter.post("/missions/:storeId", handleAddMission);
+//미션 목록 조회
+missionsRouter.get("/missions/stores/:storeId", handleListStoreMissions);
+// 유저한테 미션 도전
+missionsRouter.post("/missions/users/:userId/:missionId", handleChallengeMission);
+//내가 진행중인 미션 목록
+missionsRouter.get("/missions/users/:userId", handleListUserMissions);
 
 export default missionsRouter;
