+++ eden/week6/src/modules/reviews/reviews.routes.ts
@@ -1,9 +1,10 @@
 import { Router } from "express";
-import { handleAddReview, handleListMyReviews } from "./controllers/review.controller.js";
+import { handleAddReview, handleListMyReviews, handleListStoreReviews } from "./controllers/review.controller.js";
 
 const reviewsRouter = Router();
 
 // /api/v1/stores/:storeId/reviews
 reviewsRouter.post("/reviews/:storeId", handleAddReview);
 reviewsRouter.get("/reviews", handleListMyReviews);
+reviewsRouter.get("/reviews/:storeId", handleListStoreReviews);
 export default reviewsRouter;
