+++ eden/week6/src/modules/reviews/services/review.service.ts
@@ -1,5 +1,5 @@
-import {addReview, getReviewsByUserId} from "../repositories/review.repositoriy.js";
-import { AddReviewRequestDTO, responseFromMyReviews } from "../dtos/review.dto.js";
+import {addReview, getAllStoreReviews, getReviewsByUserId} from "../repositories/review.repositoriy.js";
+import { AddReviewRequestDTO, responseFromMyReviews, responseFromReviews, ReviewListResponse } from "../dtos/review.dto.js";
 import { getStoreById } from "../repositories/review.repositoriy.js";
 
 
@@ -12,7 +12,7 @@ export const createReview = async (storeId: number, reviewData: AddReviewRequest
     throw new Error("존재하지 않는 가게입니다.");
   }
 
-  // 가게가 존재하면 리뷰를 추가 (DTO에서 userId 추출)
+  // 가게가 존재하면 리뷰를 추가
   const newReviewId = await addReview(storeId, reviewData.userId, {
     rating: reviewData.rating,
     comment: reviewData.comment,
@@ -24,9 +24,14 @@ export const createReview = async (storeId: number, reviewData: AddReviewRequest
 
 // 내가 작성한 리뷰 목록 조회 서비스
 export const listMyReviews = async (userId: number) => {
-    //레포지토리를 호출하여 데이터베이스에서 리뷰 목록을 가져옵니다.
+    
     const reviews = await getReviewsByUserId(userId);
 
-    //가져온 데이터를 DTO를 사용해 API 응답 형식으로 변환합니다.
+    
     return responseFromMyReviews(reviews);
 }
+export const listStoreReviews = async (
+storeId: number, cursor: number): Promise<ReviewListResponse> => {
+  const reviews = await getAllStoreReviews(storeId, cursor);
+  return responseFromReviews(reviews);
+};
