+++ eden/week6/src/modules/stores/dtos/store.dtos.ts
@@ -6,9 +6,9 @@ export interface AddStoreRequestDTO {
 
 // 응답 타입을 정의 
 export interface StoreResponseDTO {
-  id: number;//가게 아이디
-  name: string;//가게 이름
-  region: string; // 예시로 지역 이름도 포함
+  id: number;
+  name: string;
+  region: string; 
   foodCategory: string;
 }
 
@@ -21,30 +21,3 @@ export const responseFromStore = (store: any) => {
   };
 };
 
-export const responseFromReviews = (
-reviews: ReviewItem[]): ReviewListResponse => {
-    const lastReview = reviews[reviews.length - 1];
-  
-    return {
-      data: reviews,
-      pagination: {
-        cursor: lastReview ? lastReview.id : null,
-      },
-    };
-  };
-export interface ReviewItem {
-  id: number;
-  comment: string;
-  store: {
-    name: string;
-  };
-  user: {
-    name: string;
-  };
-}
-export interface ReviewListResponse {
-  data: ReviewItem[];
-  pagination: {
-    cursor: number | null; // 다음 페이지를 위한 커서 (마지막 리뷰의 ID)
-  };
-}
