+++ eden/week6/src/modules/users/users.routes.ts
@@ -1,12 +1,10 @@
 import { Router } from "express";
-import { handleUserSignUp, handleChallengeMission } from "./controllers/user.controllers.js";
+import { handleUserSignUp } from "./controllers/user.controllers.js";
 
 const usersRouter = Router();
 
-// /api/v1/users/signup
+//회원가입
 usersRouter.post("/users/signup", handleUserSignUp);
 
-// /api/v1/users/:userId/missions/:missionId
-usersRouter.post("/users/:userId/missions/:missionId", handleChallengeMission);
 
 export default usersRouter;
